requirements.txt
youtube-transcript-api
google-generativeai
langchain
faiss-cpu
langchain-google-genai
python-dotenv
streamlit
